Add all current payload types to message pack/unpack implementation

+++ bee-message/src/payload/drng/beacon/collective_beacon.rs
@@ -21,7 +21,7 @@ pub struct CollectiveBeaconPayload {
 }
 
 impl CollectiveBeaconPayload {
-    pub const KIND: usize = 6;
+    pub const KIND: u32 = 6;
 
     pub fn builder() -> CollectiveBeaconPayloadBuilder {
         CollectiveBeaconPayloadBuilder::default()
